moved vision stuff to its own directory
