+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Vision/VisionThreadData.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Subsystems;
+package org.firstinspires.ftc.teamcode.Subsystems.Vision;
 
 import org.apache.commons.math3.geometry.euclidean.twod.Vector2D;
 import org.firstinspires.ftc.teamcode.DriveControl.BoundingBox;
@@ -7,7 +7,7 @@ import java.io.File;
 
 public class VisionThreadData {
     private static VisionThreadData vtd;
-    public static final File modelFile = new File(""); //TODO: Set this to the TensorFlow model file when trained.
+    public static final File modelFile = new File(""); // TODO: Set this to the TensorFlow model file when trained.
     private static Vector2D distance;
 
     public static BoundingBox theoreticalPosition;
